👌 Move example routers to sub directory.
