+++ platform/lib/platform.js
@@ -27,11 +27,13 @@ const config = require('./config.js');
 
 const WWW_PREFIX = 'www.';
 const routers = {
-  'whoAmI': require('./routers/whoAmI.js'),
-  'pages': require('./routers/pages.js'),
-  'exampleSources': require('@lib/routers/exampleSources.js'),
-  'exampleEmbeds': require('@lib/routers/exampleEmbeds.js'),
-  'static': require('./routers/static.js'),
+  'whoAmI': require('@lib/routers/whoAmI.js'),
+  'pages': require('@lib/routers/pages.js'),
+  'example': {
+    'sources': require('@lib/routers/example/sources.js'),
+    'embeds': require('@lib/routers/example/embeds.js'),
+  },
+  'static': require('@lib/routers/static.js'),
   'playground': require('../../playground/backend/'),
   'boilerplate': require('../../boilerplate/backend/'),
 };
@@ -125,8 +127,8 @@ class Platform {
 
   _registerRouters() {
     this.server.use('/who-am-i', routers.whoAmI);
-    this.server.use(routers.exampleEmbeds);
-    this.server.use(routers.exampleSources);
+    this.server.use(routers.example.embeds);
+    this.server.use(routers.example.sources);
     this.server.use(routers.static);
     this.server.use('/playground', routers.playground);
     this.server.use('/boilerplate', routers.boilerplate);
