+++ platform/lib/routers/example/sources.js
